requirements.txt
Office365-REST-Python-Client
datetime
python-dotenv
fastapi
uvicorn